Prompt config (#31)
Allow for the config to be in an additional config file `appsettings.prompts.json` to separate those into it's own file.
This is just another general config file, but it is good to have the prompts in a separate file if you working with the prompts while sharing screen och streaming. Then you don't have to show the rest of the config at the same time.
This feature also enables the config to be reloaded on change so you can update the prompts while running.